Update study-info and service-info & discount as cash

+++ database/migrations/2019_12_02_132635_create_invoices_table.php
@@ -17,7 +17,7 @@ class CreateInvoicesTable extends Migration
             $table->bigIncrements('id');
             $table->bigInteger('student_id');
             $table->double('balance');
-            $table->integer('discount');
+            $table->double('discount');
             $table->double('after_discount');
             $table->boolean('payment_status');
             $table->string('invoice_date');
